Rename client base class and interface
