+++ tests/classes/test_client.php
@@ -19,9 +19,9 @@ namespace tool_objectfs\tests;
 defined('MOODLE_INTERNAL') || die();
 
 use tool_objectfs\object_file_system;
-use tool_objectfs\client\object_client;
+use tool_objectfs\client\object_client_base;
 
-class test_client extends object_client {
+class test_client extends object_client_base {
 
     private $bucketpath;
